feat: search function and update table
