+++ frontend/admin/src/assets/js/search.js
@@ -1,0 +1,8 @@
+// search.js
+function debounce(fn, delay) {
+  let timeoutId;
+  return (...args) => {
+    clearTimeout(timeoutId);
+    timeoutId = setTimeout(() => fn(...args), delay);
+  };
+}
